feat: add table pagination
