+++ web/pages/customers/index.vue
@@ -21,6 +21,11 @@
                 :data="modelData"
                 :actions="actions"
                 :paginator-info="paginatorInfo"
+                :first-page="firstPage"
+                :prev-page="prevPage"
+                :next-page="nextPage"
+                :last-page="lastPage"
+                :number-page="numberPage"
             />
 
             <ModalCRUD
@@ -48,6 +53,7 @@
 
 <script setup lang="ts">
 import type { Headers, CrudModalField } from '~/types';
+
 import { useModelCrud } from '~/composables/useModelCrud';
 
 definePageMeta({
@@ -107,5 +113,10 @@ const {
     isLoading,
     actions,
     paginatorInfo,
+    firstPage,
+    prevPage,
+    nextPage,
+    lastPage,
+    numberPage,
 } = await useModelCrud(modelName, modelFields);
 </script>
