+++ web/pages/settings/logs.vue
@@ -21,6 +21,11 @@
                 :data="modelData"
                 :actions="actions"
                 :paginator-info="paginatorInfo"
+                :first-page="firstPage"
+                :prev-page="prevPage"
+                :next-page="nextPage"
+                :last-page="lastPage"
+                :number-page="numberPage"
             />
         </main>
     </div>
@@ -28,6 +33,7 @@
 
 <script setup lang="ts">
 import type { Headers, CrudModalField } from '~/types';
+
 import { useModelCrud } from '~/composables/useModelCrud';
 
 definePageMeta({
@@ -57,6 +63,11 @@ const {
     isLoading,
     actions,
     paginatorInfo,
+    firstPage,
+    prevPage,
+    nextPage,
+    lastPage,
+    numberPage,
 } = await useModelCrud(modelName, modelFields);
 
 onMounted(async () => {
